refactor: move form animation to project form for reuse
